# payments/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Tickets Service HTTP API Docs",
        "title": "Tickets Service",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "fiber@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
        "/tickets": {
            "get": {
                "description": "Fetches all ticket data.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Get All Tickets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.TicketResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new ticket.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Create Ticket",
                "parameters": [
                    {
                        "description": "Ticket Create Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateTicketRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.TicketResponse"
                        }
                    }
                }
            }
        },
        "/tickets/{id}": {
            "get": {
                "description": "Fetches a ticket by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Get A Ticket",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ticket ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.TicketResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates a ticket data by id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Update A Ticket",
                "parameters": [
                    {
                        "description": "Ticket Update Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateTicketRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Ticket ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.TicketResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.CreateTicketRequest": {
            "type": "object",
            "required": [
                "price",
                "title",
                "user_id"
            ],
            "properties": {
                "price": {
                    "type": "integer"
                },
                "title": {
                    "type": "string",
                    "maxLength": 100
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "model.TicketResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "order_id": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "model.UpdateTicketRequest": {
            "type": "object",
            "required": [
                "id",
                "price",
                "title",
                "user_id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "order_id": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "title": {
                    "type": "string",
                    "maxLength": 100
                },
                "user_id": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Session": {
            "type": "apiKey",
            "name": "Cookie",
            "in": "header"
        }
    }
}